Split util functions, dateutils to arrows
